python_service/requirements_pymupdf_pro.txt
# FastAPI 相关
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 文档处理 - PyMuPDF Pro 统一方案
pymupdfpro>=1.26.0
pymupdf4llm>=0.0.21
llama-index>=0.9.0

# 传统方案 (作为回退)
# python-docx
# openpyxl
# python-pptx

# LangChain 相关
langchain==0.1.0
langchain-community==0.0.10
langchain-elasticsearch==0.0.1

# 向量化和搜索
sentence-transformers==2.2.2
elasticsearch==8.11.0

# 其他依赖
python-ldap==3.4.3
requests==2.31.0
numpy==1.24.3 